requirements.txt
requests>=2.28.0
openai>=1.0.0
pandas>=1.5.0
python-dotenv>=0.19.0
anthropic>=0.7.0
google-generativeai>=0.3.0 